remove useless words in file names and organize them properly
