+++ lib/components/subject_management/subject_configs/color.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:timetable/components/subject_management/subject_configs/colors_config_screens/colors_screen.dart';
+import 'package:timetable/components/subject_management/subject_configs/colors_screens/colors.dart';
 import 'package:timetable/components/widgets/list_item_group.dart';
 
 /// Color configuration part of the Subject creation screen.
