+++ lib/components/subject_management/subject_configs/day.dart
