+++ lib/components/subject_management/subject_configs/rotation_week.dart
