+++ lib/components/subject_management/subject_screen.dart
@@ -2,10 +2,10 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:timetable/components/subject_management/subject_configs/colors_config.dart';
+import 'package:timetable/components/subject_management/subject_configs/color.dart';
 import 'package:timetable/components/subject_management/subject_configs/day_time_week_tb_config.dart';
-import 'package:timetable/components/subject_management/subject_configs/label_location_config.dart';
-import 'package:timetable/components/subject_management/subject_configs/note_tile.dart';
+import 'package:timetable/components/subject_management/subject_configs/label_location.dart';
+import 'package:timetable/components/subject_management/subject_configs/note.dart';
 import 'package:timetable/constants/days.dart';
 import 'package:timetable/constants/rotation_weeks.dart';
 import 'package:timetable/db/database.dart';
