+++ lib/components/timetable_views/grid.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:timetable/components/widgets/grid.dart';
-import 'package:timetable/components/widgets/grid_view_overlapping_subjects_builder.dart';
+import 'package:timetable/components/widgets/grid_view/grid.dart';
+import 'package:timetable/components/widgets/grid_view/grid_view_overlapping_subjects_builder.dart';
 import 'package:timetable/components/widgets/time_column.dart';
 import 'package:timetable/constants/custom_times.dart';
 import 'package:timetable/constants/grid_properties.dart';
@@ -10,9 +10,9 @@ import 'package:timetable/constants/rotation_weeks.dart';
 import 'package:timetable/db/database.dart';
 import 'package:timetable/helpers/timetables.dart';
 import 'package:timetable/provider/overlapping_subjects.dart';
-import 'package:timetable/components/widgets/grid_view_subject_builder.dart';
-import 'package:timetable/components/widgets/grid_view_subject_container_builder.dart';
-import 'package:timetable/components/widgets/tile.dart';
+import 'package:timetable/components/widgets/grid_view/grid_view_subject_builder.dart';
+import 'package:timetable/components/widgets/grid_view/grid_view_subject_container_builder.dart';
+import 'package:timetable/components/widgets/grid_view/tile.dart';
 import 'package:timetable/helpers/rotation_weeks.dart';
 import 'package:timetable/provider/settings.dart';
 import 'package:timetable/provider/timetables.dart';
