+++ lib/components/widgets/bottom_sheets/days_modal_bottom_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:timetable/components/widgets/subject_data_bottom_sheet.dart';
+import 'package:timetable/components/widgets/bottom_sheets/subject_data_bottom_sheet.dart';
 import 'package:timetable/constants/days.dart';
 
 /// Bottom Sheet Modal Widget used to select a subject's day.
