+++ lib/components/widgets/bottom_sheets/rotation_week_modal_bottom_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:timetable/components/widgets/subject_data_bottom_sheet.dart';
+import 'package:timetable/components/widgets/bottom_sheets/subject_data_bottom_sheet.dart';
 import 'package:timetable/constants/rotation_weeks.dart';
 import 'package:timetable/helpers/rotation_weeks.dart';
