+++ lib/components/widgets/day_view/days_bar.dart
