+++ lib/screens/timetable.dart
@@ -2,10 +2,10 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:timetable/components/timetable_views/day_view.dart';
-import 'package:timetable/components/timetable_views/grid_view.dart';
-import 'package:timetable/components/widgets/days_bar.dart';
-import 'package:timetable/components/widgets/grid_day_views_toggle.dart';
+import 'package:timetable/components/timetable_views/day.dart';
+import 'package:timetable/components/timetable_views/grid.dart';
+import 'package:timetable/components/widgets/day_view/days_bar.dart';
+import 'package:timetable/components/widgets/grid_view/grid_day_views_toggle.dart';
 import 'package:timetable/components/widgets/navigation_bar_toggle.dart';
 import 'package:timetable/components/widgets/rotation_week_toggle.dart';
 import 'package:timetable/components/widgets/timetable_toggle.dart';
